refactor(aether-actor): collapse or rename the stuttering actor submodule
Rename `crates/aether-actor/src/actor/` → `src/model/` to eliminate
the `aether_actor::actor::*` stutter. The module rename is a `git mv`;
the three crate-root flat re-exports are repointed from `actor::` to
`model::` so all public type paths (`aether_actor::Addressable`,
`aether_actor::OutboundReply`, etc.) are unchanged for downstream
consumers. Internal `crate::actor::` / `crate::actor::ctx::*` paths
in `wasm/ctx.rs`, `wasm/mailbox.rs`, `wasm/inline/mod.rs`, and the
two bridge modules are rewritten to `crate::model::`. External
`aether_actor::actor::` module-path sites in `aether-capabilities` and
`aether-substrate` are rewritten to `model::` or collapsed to the
already-promoted crate-root flat re-export. The ~47
`use aether_actor::actor;` macro-import sites are untouched (macro
namespace, separate from the module rename).

Closes #2419
