+++ crates/aether-actor/src/model/slot.rs
