+++ crates/aether-actor/src/wasm/inline/mod.rs
@@ -599,8 +599,8 @@ where
 mod tests {
     use super::{InlineRegistry, RouteDecision, drain_cluster_queue, membrane_dispatch};
     use crate::WasmCtx;
-    use crate::actor::ctx::OutboundReply;
     use crate::mail::{Mail, PriorState};
+    use crate::model::ctx::OutboundReply;
     use crate::wasm::ErasedWasmActor;
     use aether_data::MailboxId;
     use alloc::boxed::Box;
