+++ crates/aether-capabilities/src/lifecycle/runtime/mod.rs
@@ -35,7 +35,7 @@ use aether_kinds::{
 };
 
 pub use aether_actor::Manual;
-pub use aether_actor::actor::ctx::OutboundReply;
+pub use aether_actor::OutboundReply;
 pub use aether_data::{Kind, KindId, MailboxId as DataMailboxId, mailbox_id_from_name};
 pub use aether_kinds::LifecycleAdvanceComplete;
 pub use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx};
